services/rag/requirements.txt
fastapi>=0.115.0,<1
uvicorn[standard]>=0.32.0,<1
httpx>=0.27.0,<1
pydantic-settings>=2.0.0,<3
PyJWT[crypto]>=2.8.0,<3
